statistics: add statslock test
