+++ statistics/handle/handletest/BUILD.bazel
@@ -9,7 +9,7 @@ go_test(
     ],
     flaky = True,
     race = "on",
-    shard_count = 47,
+    shard_count = 45,
     deps = [
         "//config",
         "//domain",
